Backend/requirements.txt
flask
Flask[async]
supabase
python-dotenv
flask-cors
